% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximate_discretized_score.R
\name{approximate_discretized_score}
\alias{approximate_discretized_score}
\title{MCMC estimator of score function at a discretization level}
\usage{
approximate_discretized_score(
  model,
  theta,
  discretization,
  observations,
  nparticles,
  resampling_threshold = 1,
  initialization = "particlefilter",
  algorithm = "CPF",
  max_iterations
)
}
\arguments{
\item{model}{a list representing a hidden Markov model, e.g. \code{\link{hmm_ornstein_uhlenbeck}}}

\item{theta}{a vector of parameters as input to model functions}

\item{discretization}{list containing stepsize, nsteps, statelength and obstimes}

\item{observations}{a matrix of observations of size terminal_time x ydimension}

\item{nparticles}{number of particles}

\item{resampling_threshold}{ESS proportion below which resampling is triggered (always resample at observation times by default)}

\item{initialization}{choice of distribution to initialize chains, such as \code{dynamics} or the default \code{particlefilter}}

\item{algorithm}{character specifying type of algorithm desired, i.e. 
\code{\link{CPF}} for conditional particle filter, 
\code{\link{CASPF}} for conditional ancestor sampling particle filter,
\code{\link{CBSPF}} for conditional backward sampling particle filter}

\item{max_iterations}{number of MCMC iterations}
}
\value{
a list with objects such as: 
\code{mcmcestimator} is the MCMC estimator of the discretized score; 
\code{cost} is the cost of the algorithm; 
\code{elapsedtime} is the time taken by the algorithm.
}
\description{
Computes MCMC estimator of the time-discretized score function.
}
