% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBSPF.R
\name{CBSPF}
\alias{CBSPF}
\title{Conditional Backward Sampling Particle Filter}
\usage{
CBSPF(
  model,
  theta,
  discretization,
  observations,
  nparticles,
  resampling_threshold = 1,
  ref_trajectory
)
}
\arguments{
\item{model}{a list representing a hidden Markov model, e.g. \code{\link{hmm_ornstein_uhlenbeck}}}

\item{theta}{a vector of parameters as input to model functions}

\item{discretization}{list containing stepsize, nsteps, statelength and obstimes}

\item{observations}{a matrix of observations, of size nobservations x ydimension}

\item{nparticles}{number of particles}

\item{resampling_threshold}{ESS proportion below which resampling is triggered (always resample at observation times by default)}

\item{ref_trajectory}{a matrix of reference trajectory, of size xdimension x statelength}
}
\value{
a matrix containing a new trajectory of size xdimension x statelength.
}
\description{
Runs a conditional particle filter with backward sampling (Whiteley, 2010).
}
