% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coupled2_CASPF.R
\name{coupled2_CASPF}
\alias{coupled2_CASPF}
\title{2-way Coupled Conditional Ancestor Sampling Particle Filter}
\usage{
coupled2_CASPF(
  model,
  theta,
  discretization,
  observations,
  nparticles,
  resampling_threshold = 1,
  coupled_resampling,
  ref_trajectory1,
  ref_trajectory2,
  treestorage = FALSE
)
}
\arguments{
\item{model}{a list representing a hidden Markov model, e.g. \code{\link{hmm_ornstein_uhlenbeck}}}

\item{theta}{a vector of parameters as input to model functions}

\item{discretization}{list containing stepsize, nsteps, statelength and obstimes}

\item{observations}{a matrix of observations, of size nobservations x ydimension}

\item{nparticles}{number of particles}

\item{resampling_threshold}{ESS proportion below which resampling is triggered (always resample at observation times by default)}

\item{coupled_resampling}{a 2-way coupled resampling scheme, such as \code{\link{coupled2_maximal_independent_residuals}}}

\item{ref_trajectory1}{a matrix of first reference trajectory, of size xdimension x statelength}

\item{ref_trajectory2}{a matrix of second reference trajectory, of size xdimension x statelength}

\item{treestorage}{logical specifying tree storage of Jacob, Murray and Rubenthaler (2013);
if missing, this function store all states and ancestors}
}
\value{
a pair of new trajectories stored as matrices of size xdimension x statelength.
}
\description{
Runs two coupled conditional particle filters (at each discretization level).
}
