% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coupled2_CPF.R
\name{coupled2_CPF}
\alias{coupled2_CPF}
\title{2-way Coupled Conditional Particle Filter}
\usage{
coupled2_CPF(
  model,
  theta,
  level,
  observations,
  nparticles,
  coupled_resampling,
  ref_trajectory1,
  ref_trajectory2
)
}
\arguments{
\item{model}{a list representing a hidden Markov model, e.g. \code{\link{hmm_ornstein_uhlenbeck}}}

\item{theta}{a vector of parameters as input to model functions}

\item{level}{level of the time discretization}

\item{observations}{a matrix of observations of size terminal_time x ydimension}

\item{nparticles}{number of particles}

\item{coupled_resampling}{a 2-way coupled resampling scheme, such as \code{\link{coupled2_maximal_coupled_residuals}}}

\item{ref_trajectory1}{a matrix of first reference trajectory, of size xdimension x statelength}

\item{ref_trajectory2}{a matrix of second reference trajectory, of size xdimension x statelength}
}
\value{
a pair of new trajectories stored as matrices of size xdimension x statelength
}
\description{
Runs two coupled conditional particle filters (at each discretization level)
}
