% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coupled2_resampling.R
\name{coupled2_maximal_coupled_residuals}
\alias{coupled2_maximal_coupled_residuals}
\title{(2-way) Maximally coupled multinomial resampling with coupled residuals 
using common uniform random variables}
\usage{
coupled2_maximal_coupled_residuals(normweights1, normweights2, ndraws, rand)
}
\arguments{
\item{normweights1}{normalized weights of first particle system}

\item{normweights2}{normalized weights of second particle system}

\item{ndraws}{number of samples from multinomial distribution}

\item{rand}{common uniform random variables used for the coupled resampling}
}
\value{
A matrix of ancestor indexes of size ndraws x 2
}
\description{
This function performs (2-way) maximally coupled multinomial resampling 
with coupled residuals using common uniform random variables on each particle system
}
