% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coupled2_resampling.R
\name{coupled2_resampling}
\alias{coupled2_resampling}
\title{(2-way) Coupled multinomial resampling using common uniforms random variables}
\usage{
coupled2_resampling(normweights1, normweights2, ndraws, rand)
}
\arguments{
\item{normweights1}{normalized weights of first particle system}

\item{normweights2}{normalized weights of second particle system}

\item{ndraws}{number of samples from multinomial distribution}

\item{rand}{common uniform random variables used for the coupled resampling}
}
\value{
A matrix of ancestor indexes of size ndraws x 2
}
\description{
This function performs (2-way) coupled multinomial resampling 
on each particle system using common uniform random variables
}
