% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coupled4_CBSPF.R
\name{coupled4_CBSPF}
\alias{coupled4_CBSPF}
\title{4-way Coupled Conditional Particle Filter with Backward Sampling}
\usage{
coupled4_CBSPF(
  model,
  theta,
  discretization,
  observations,
  nparticles,
  resampling_threshold,
  coupled_resampling,
  ref_trajectory_coarse1,
  ref_trajectory_coarse2,
  ref_trajectory_fine1,
  ref_trajectory_fine2
)
}
\arguments{
\item{model}{a list representing a hidden Markov model, e.g. \code{\link{hmm_ornstein_uhlenbeck}}}

\item{theta}{a vector of parameters as input to model functions}

\item{discretization}{lists containing stepsize, nsteps, statelength, obstimes for fine and coarse levels, 
and coarsetimes of length statelength_fine indexing time steps of coarse level}

\item{observations}{a matrix of observations, of size nobservations x ydimension}

\item{nparticles}{number of particles}

\item{resampling_threshold}{ESS proportion below which resampling is triggered (always resample at observation times by default)}

\item{coupled_resampling}{a 4-way coupled resampling scheme, such as \code{\link{coupled4_maximalchains_maximallevels_independent_residuals}}}

\item{ref_trajectory_coarse1}{a matrix of first reference trajectory for coarser discretization level, of size xdimension x statelength_coarse}

\item{ref_trajectory_coarse2}{a matrix of second reference trajectory for coarser discretization level, of size xdimension x statelength_coarse}

\item{ref_trajectory_fine1}{a matrix of first reference trajectory for finer discretization level, of size xdimension x statelength_fine}

\item{ref_trajectory_fine2}{a matrix of second reference trajectory for finer discretization level, of size xdimension x statelength_fine}
}
\value{
four new trajectories stored as matrices of size xdimension x statelength_coarse/fine.
}
\description{
Runs four coupled conditional particle filters (two at each discretization level).
}
