% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coupled4_resampling.R
\name{coupled4_maximal_independent_residuals}
\alias{coupled4_maximal_independent_residuals}
\title{(4-way) Maximally coupled multinomial resampling with independent residuals}
\usage{
coupled4_maximal_independent_residuals(
  normweights1,
  normweights2,
  normweights3,
  normweights4,
  ndraws,
  rand
)
}
\arguments{
\item{normweights1}{normalized weights of first particle system}

\item{normweights2}{normalized weights of second particle system}

\item{normweights3}{normalized weights of third particle system}

\item{normweights4}{normalized weights of fourth particle system}

\item{ndraws}{number of samples from multinomial distribution}

\item{rand}{common uniform random variables used for the coupled resampling}
}
\value{
A matrix of ancestor indexes of size ndraws x 4s
}
\description{
This function performs (4-way) maximally coupled multinomial resampling 
with independent residuals on each particle system
}
