% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coupled4_resampling.R
\name{coupled4_maximallevels_maximalchains_coupled_residuals}
\alias{coupled4_maximallevels_maximalchains_coupled_residuals}
\title{(4-way) coupled multinomial resampling by maximally coupling pairs of indexes 
across levels and maximally coupling chains on each level with 
coupled residuals using common uniform random variables}
\usage{
coupled4_maximallevels_maximalchains_coupled_residuals(
  normweights_coarse1,
  normweights_coarse2,
  normweights_fine1,
  normweights_fine2,
  ndraws,
  rand
)
}
\arguments{
\item{normweights_coarse1}{normalized weights of first particle system on coarse level}

\item{normweights_coarse2}{normalized weights of second particle system on coarse level}

\item{normweights_fine1}{normalized weights of first particle system on fine level}

\item{normweights_fine2}{normalized weights of second particle system on fine level}

\item{ndraws}{number of samples from multinomial distribution}

\item{rand}{common uniform random variables used for the coupled resampling}
}
\value{
A matrix of ancestor indexes of size ndraws x 4
}
\description{
This function performs (4-way) coupled multinomial resampling by maximally coupling pairs 
of indexes across levels and maximally coupling chains on each level with with coupled residuals 
using common uniform random variables on each particle system
}
