% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{exp_transform}
\alias{exp_transform}
\title{Exponential transformation}
\usage{
exp_transform(model, log_theta)
}
\arguments{
\item{model}{a list representing a hidden Markov model, e.g. \code{\link{hmm_ornstein_uhlenbeck}}}

\item{log_theta}{a vector of parameters}
}
\value{
a vector of transformed parameters
}
\description{
Computes exponential transformation for parameters with positivity constraints.
}
