% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_path.R
\name{get_path_bs}
\alias{get_path_bs}
\title{Get a path from the output of a particle filter using backward sampling (Whiteley, 2010)}
\usage{
get_path_bs(
  model,
  discretization,
  xtrajectory,
  ancestries,
  ancestor,
  store_logweights
)
}
\arguments{
\item{model}{a list representing a hidden Markov model, e.g. \code{\link{hmm_ornstein_uhlenbeck}}}

\item{discretization}{list containing stepsize, nsteps, statelength and obstimes}

\item{xtrajectory}{an array of particle states, of size statelength x xdimension x nparticles}

\item{ancestries}{a matrix of ancestor indexes, of size statelength x nparticles}

\item{ancestor}{an ancestor index at the terminal step}

\item{store_logweights}{a matrix of log-weights, of size nobservations x nparticles}
}
\value{
a matrix containing a new trajectory of size xdimension x statelength
}
\description{
Get a path from the output of a particle filter by sampling an ancestry
}
