% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm_logistic_diffusion.R
\name{hmm_logistic_diffusion}
\alias{hmm_logistic_diffusion}
\title{Construct hidden markov model obtained by discretizing logistic diffusion process}
\usage{
hmm_logistic_diffusion(times, sigma0)
}
\arguments{
\item{times}{vector specifying observation times}

\item{sigma0}{diffusivity parameter}
}
\value{
A list
}
\description{
This function returns a list with objects such as
* xdimension is the dimension of the latent process
* ydimension is the dimension of the observation process
* theta_dimension is the dimension of the parameter space
* construct_discretization outputs a list containing stepsize, nsteps, statelength and obstimes
* construct_successive_discretization outputs lists containing stepsize, nsteps, statelength, obstimes for fine and coarse levels, 
and coarsetimes of length statelength_fine indexing time steps of coarse level
* sigma is the diffusion coefficient of the process
* rinit to sample from the initial distribution
* rtransition to sample from the Markov transition
* dtransition to evaluate the transition density
* dmeasurement to evaluate the measurement density
* functional is the smoothing function to compute gradient of log-likelihood
}
