% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm_logistic_diffusion_full.R
\name{hmm_logistic_diffusion_full}
\alias{hmm_logistic_diffusion_full}
\title{Construct hidden Markov model obtained by discretizing Lamperti transformed logistic diffusion process}
\usage{
hmm_logistic_diffusion_full(times)
}
\arguments{
\item{times}{vector specifying observation times}
}
\value{
a list with objects such as: 
\code{xdimension} is the dimension of the latent process; 
\code{ydimension} is the dimension of the observation process; 
\code{theta_dimension} is the dimension of the parameter space; 
\code{theta_names} is a vector of characters to enumerate the parameters;
\code{theta_positivity} is a vector logicals to index parameters with positivity constraints;
\code{construct_discretization} outputs a list containing stepsize, nsteps, statelength and obstimes; 
\code{construct_successive_discretization} outputs lists containing stepsize, nsteps, statelength, obstimes for fine and coarse levels, 
and coarsetimes of length statelength_fine indexing time steps of coarse level;
\code{sigma} is the diffusion coefficient of the process; 
\code{rinit} to sample from the initial distribution; 
\code{rtransition} to sample from the Markov transition; 
\code{dtransition} to evaluate the transition density; 
\code{dmeasurement} to evaluate the measurement density; 
\code{functional} is the smoothing function to compute gradient of log-likelihood.
}
\description{
Define a hidden Markov model obtained by discretizing Lamperti transformed logistic diffusion process.
}
