% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm_neuroscience_diffusion.R
\name{hmm_neuroscience_diffusion}
\alias{hmm_neuroscience_diffusion}
\title{Construct hidden markov model obtained by discretizing neuroscience diffusion}
\usage{
hmm_neuroscience_diffusion(
  spiketimes1,
  spiketimes2,
  level_observation,
  terminal_time
)
}
\arguments{
\item{spiketimes1}{vector specifying cell spike times for the first grid cell}

\item{spiketimes2}{vector specifying cell spike times for the second grid cell}
}
\value{
A list
}
\description{
This function returns a list with objects such as
* xdimension is the dimension of the latent process, 
* ydimension is the dimension of the observation process, 
* theta_dimension is the dimension of the parameter space, 
* compute_observations compute observation counts given spike times and discretization level, 
* construct_discretization outputs a list containing stepsize, nsteps, statelength and obstimes, 
* construct_successive_discretization outputs lists containing stepsize, nsteps, statelength, obstimes for fine and coarse levels, 
and coarsetimes of length statelength_fine indexing time steps of coarse level
* sigma is the diffusion coefficient of the process, 
* rinit to sample from the initial distribution, 
* rtransition to sample from the Markov transition, 
* dmeasurement to evaluate the measurement density, 
* functional is the smoothing function to compute gradient of log-likelihood.
}
