% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm_ornstein_uhlenbeck.R
\name{hmm_ornstein_uhlenbeck}
\alias{hmm_ornstein_uhlenbeck}
\title{Construct hidden markov model obtained by discretizing Ornstein-Uhlenbeck process}
\usage{
hmm_ornstein_uhlenbeck(terminal_time)
}
\arguments{
\item{terminal_time}{number specifying time interval of interest}
}
\value{
A list
}
\description{
This function returns a list with objects such as
* xdimension is the dimension of the latent process
* ydimension is the dimension of the observation process
* theta_dimension is the dimension of the parameter space
* statelength is the length of the discretized process at each level
* obstimes is a vector of logicals indexing the observation times
* sigma is the diffusion coefficient of the process
* rinit to sample from the initial distribution
* rtransition to sample from the Markov transition
* dtransition to evaluate the transition density
* dmeasurement to evaluate the measurement density
* functional is the smoothing function to compute gradient of log-likelihood
}
