% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{log_transform}
\alias{log_transform}
\title{Logarithmic transformation}
\usage{
log_transform(model, theta)
}
\arguments{
\item{model}{a list representing a hidden Markov model, e.g. \code{\link{hmm_ornstein_uhlenbeck}}}

\item{theta}{a vector of parameters}
}
\value{
a vector of transformed parameters
}
\description{
Computes logarithmic transformation for parameters with positivity constraints.
}
