% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_SDE.R
\name{simulate_SDE}
\alias{simulate_SDE}
\title{Simulate time-discretized process following a stochastic differential equation}
\usage{
simulate_SDE(model, theta, discretization)
}
\arguments{
\item{model}{a list representing a hidden Markov model, e.g. \code{\link{hmm_ornstein_uhlenbeck}}}

\item{theta}{a vector of parameters as input to model functions}

\item{discretization}{list containing stepsize, nsteps and statelength}
}
\value{
a matrix containing a new trajectory of size xdimension x statelength.
}
\description{
Simulate a trajectory following a stochastic differential equation using Euler-Maruyama discretization.
}
