% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unbiased_estimators.R
\name{systematic_estimator}
\alias{systematic_estimator}
\title{Unbiased estimators of the score function based on systematic sampling}
\usage{
systematic_estimator(
  model,
  theta,
  observations,
  nparticles,
  resampling_threshold = 1,
  coupled2_resampling,
  coupled4_resampling,
  initialization = "particlefilter",
  algorithm = "CPF",
  k = 0,
  m = 1,
  level_distribution,
  maxnrepeats
)
}
\arguments{
\item{model}{a list representing a hidden Markov model, e.g. \code{\link{hmm_ornstein_uhlenbeck}}}

\item{theta}{a vector of parameters as input to model functions}

\item{observations}{a matrix of observations of size nobservations x ydimension}

\item{nparticles}{number of particles}

\item{resampling_threshold}{ESS proportion below which resampling is triggered (always resample at observation times by default)}

\item{coupled2_resampling}{a 2-marginal coupled resampling scheme, such as \code{\link{coupled2_maximal_independent_residuals}}}

\item{coupled4_resampling}{a 4-marginal coupled resampling scheme, such as \code{\link{coupled4_maximalchains_maximallevels_independent_residuals}}}

\item{initialization}{choice of distribution to initialize chains, such as \code{dynamics} or the default \code{particlefilter}}

\item{algorithm}{character specifying type of algorithm desired, i.e. 
\code{\link{CPF}} for conditional particle filter, 
\code{\link{CASPF}} for conditional ancestor sampling particle filter,
\code{\link{CBSPF}} for conditional backward sampling particle filter}

\item{k}{iteration at which to start averaging (default to 0)}

\item{m}{iteration at which to stop averaging (default to 1)}

\item{level_distribution}{list containing mass_function and tail_function that specify the distribution of levels, 
e.g. by calling \code{\link{compute_level_distribution}}}

\item{maxnrepeats}{maximum number of replicates to be stratified}
}
\value{
a list with objects such as: 
\code{nlevel_repeats} is a matrix of the number of repetitions on each level;
\code{unbiasedestimator} is a matrix of unbiased estimators of the gradient of the log-likelihood;
\code{cost} is a vector of cost to compute the unbiased estimators;
\code{elapsedtime} is the time taken to compute the estimators.
}
\description{
Estimates the score function using the unbiased stratified estimators of Vihola (2018)
}
