% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unbiased_scores.R
\name{unbiased_discretized_score}
\alias{unbiased_discretized_score}
\title{Unbiased estimator of score function at a discretization level}
\usage{
unbiased_discretized_score(
  model,
  theta,
  discretization,
  observations,
  nparticles,
  resampling_threshold = 1,
  coupled_resampling,
  initialization = "particlefilter",
  algorithm = "CPF",
  k = 0,
  m = 1,
  max_iterations = Inf
)
}
\arguments{
\item{model}{a list representing a hidden Markov model, e.g. \code{\link{hmm_ornstein_uhlenbeck}}}

\item{theta}{a vector of parameters as input to model functions}

\item{discretization}{list containing stepsize, nsteps, statelength and obstimes}

\item{observations}{a matrix of observations of size terminal_time x ydimension}

\item{nparticles}{number of particles}

\item{resampling_threshold}{ESS proportion below which resampling is triggered (always resample at observation times by default)}

\item{coupled_resampling}{a 2-marginal coupled resampling scheme, such as \code{\link{coupled2_maximal_independent_residuals}}}

\item{initialization}{choice of distribution to initialize chains, such as \code{dynamics} or the default \code{particlefilter}}

\item{algorithm}{character specifying type of algorithm desired, i.e. 
\code{\link{CPF}} for conditional particle filter, 
\code{\link{CASPF}} for conditional ancestor sampling particle filter,
\code{\link{CBSPF}} for conditional backward sampling particle filter}

\item{k}{iteration at which to start averaging (default to 0)}

\item{m}{iteration at which to stop averaging (default to 1)}

\item{max_iterations}{iteration at which to stop the while loop (default to infinity)}
}
\value{
a list with objects such as 
mcmcestimator is the MCMC estimator of the discretized score 
unbiasedestimator is an unbiased estimator of the discretized score 
meetingtime is the meeting time of the two chains at level
iteration is the number of iterations taken
finished indicates if the algorithm has completed successfully
cost is the cost of the algorithm in units of the CPF kernel at current discretization level
elapsedtime is the time taken by the algorithm
}
\description{
Computes unbiased estimator of the time-discretized score function
}
