% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unbiased_estimators.R
\name{unbiased_gradient}
\alias{unbiased_gradient}
\title{Unbiased estimator of the gradient of the log-likelihood at a discretization level}
\usage{
unbiased_gradient(
  model,
  theta,
  discretization,
  observations,
  nparticles,
  coupled_resampling,
  k = 0,
  m = 1,
  max_iterations = Inf
)
}
\arguments{
\item{model}{a list representing a hidden Markov model, e.g. \code{\link{hmm_ornstein_uhlenbeck}}}

\item{theta}{a vector of parameters as input to model functions}

\item{discretization}{list containing stepsize, nsteps, statelength and obstimes}

\item{observations}{a matrix of observations of size terminal_time x ydimension}

\item{nparticles}{number of particles}

\item{coupled_resampling}{a 2-way coupled resampling scheme, such as \code{\link{coupled2_maximal_coupled_residuals}}}

\item{k}{iteration at which to start averaging (default to 0)}

\item{m}{iteration at which to stop averaging (default to 1)}

\item{max_niterations}{iteration at which to stop the while loop (default to infinity)}
}
\value{
a list with objects such as 
mcmcestimator is the MCMC estimator of the gradient at level
unbiasedestimator is an unbiased estimator of the gradient at level
meetingtime is the meeting time of the two chains at level
iteration is the number of iterations taken
finished indicates if the algorithm has completed successfully
}
\description{
Estimates the expectation of a functional with respect to the smoothing distribution at a discretization level
}
