% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unbiased_scores.R
\name{unbiased_score_increment}
\alias{unbiased_score_increment}
\title{Unbiased estimator of score increment between two successive discretization levels}
\usage{
unbiased_score_increment(
  model,
  theta,
  discretization,
  observations,
  nparticles,
  resampling_threshold = 1,
  coupled2_resampling,
  coupled4_resampling,
  initialization = "particlefilter",
  algorithm = "CPF",
  k = 0,
  m = 1,
  max_iterations = Inf
)
}
\arguments{
\item{model}{a list representing a hidden Markov model, e.g. \code{\link{hmm_ornstein_uhlenbeck}}}

\item{theta}{a vector of parameters as input to model functions}

\item{discretization}{lists containing stepsize, nsteps, statelength, obstimes for fine and coarse levels, 
and coarsetimes of length statelength_fine indexing time steps of coarse level}

\item{observations}{a matrix of observations of size nobservations x ydimension}

\item{nparticles}{number of particles}

\item{resampling_threshold}{ESS proportion below which resampling is triggered (always resample at observation times by default)}

\item{coupled2_resampling}{a 2-marginal coupled resampling scheme, such as \code{\link{coupled2_maximal_independent_residuals}}}

\item{coupled4_resampling}{a 4-marginal coupled resampling scheme, such as \code{\link{coupled4_maximalchains_maximallevels_independent_residuals}}}

\item{initialization}{choice of distribution to initialize chains, such as \code{dynamics} or the default \code{particlefilter}}

\item{algorithm}{character specifying type of algorithm desired, i.e. 
\code{\link{CPF}} for conditional particle filter, 
\code{\link{CASPF}} for conditional ancestor sampling particle filter,
\code{\link{CBSPF}} for conditional backward sampling particle filter}

\item{k}{iteration at which to start averaging (default to 0)}

\item{m}{iteration at which to stop averaging (default to 1)}

\item{max_iterations}{iteration at which to stop the while loop (default to infinity)}
}
\value{
a list with objects such as: 
\code{mcmcestimator_coarse} is the MCMC estimator of the score at level-1; 
\code{mcmcestimator_fine} is the MCMC estimator of the score at level; 
\code{unbiasedestimator_coarse} is an unbiased estimator of the score at level-1; 
\code{unbiasedestimator_fine} is an unbiased estimator of the score at level; 
\code{mcmcestimator} is the MCMC estimator of the score increment between the two discretization levels; 
\code{unbiasedestimator} is an unbiased estimator of the score increment between the two discretization levels; 
\code{meetingtime_coarse} is the meeting time of the two chains at level-1; 
\code{meetingtime_fine} is the meeting time of the two chains at level;
\code{iteration} is the number of iterations taken; 
\code{finished} indicates if the algorithm has completed successfully; 
\code{cost_coarse} is the cost of the algorithm in units of the CPF kernel at coarse discretization level; 
\code{cost_fine} is the cost of the algorithm in units of the CPF kernel at fine discretization level; 
\code{elapsedtime} is the time taken by the algorithm.
}
\description{
Computes unbiased estimator of the difference of the time-discretized score functions between two successive discretization levels.
}
