\name{UnbiasedGradients-package}
\alias{UnbiasedGradients-package}
\alias{UnbiasedGradients}
\docType{package}
\title{
\packageTitle{UnbiasedGradients}
}
\description{
\packageDescription{UnbiasedGradients}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{UnbiasedGradients}
\packageIndices{UnbiasedGradients}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{UnbiasedGradients}

Maintainer: \packageMaintainer{UnbiasedGradients}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
