% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnbiasedScore.R
\docType{package}
\name{UnbiasedScore}
\alias{UnbiasedScore}
\title{UnbiasedScore package}
\description{
Unbiased estimators of the score function
}
\details{
This R package provides unbiased estimators of the score function, i.e. the gradient of the log-likelihood for a class of partially observed diffusions.
}
\author{
Jeremy Heng <jeremyhengjm@gmail.com>
}
\keyword{package}
