% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unbiased_estimators.R
\name{compute_level_distribution}
\alias{compute_level_distribution}
\title{Compute distribution of levels}
\usage{
compute_level_distribution(
  model,
  minimum_level,
  maximum_level = 16,
  regular_stepsizes = TRUE
)
}
\arguments{
\item{model}{a list representing a hidden Markov model, e.g. \code{\link{hmm_ornstein_uhlenbeck}}}

\item{minimum_level}{coarsest discretization level}

\item{maximum_level}{finest discretization level}
}
\value{
a list with objects such as: 
\code{support} of the level distribution;
\code{mass_function} is the probability mass function of the level distribution;
\code{tail_function} is the tail probability function of the level distribution.
}
\description{
Compute distribution of levels using results from Section 4 of Rhee and Glynn (2015).
}
