% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coupled4_resampling.R
\name{coupled4_maximalchains_maximallevels_coupled_residuals}
\alias{coupled4_maximalchains_maximallevels_coupled_residuals}
\title{(4-way) coupled multinomial resampling by maximally coupling chains and maximally coupling
across levels with coupled residuals using common uniform random variables}
\usage{
coupled4_maximalchains_maximallevels_coupled_residuals(
  normweights1,
  normweights2,
  normweights3,
  normweights4,
  ndraws,
  rand
)
}
\arguments{
\item{normweights1}{normalized weights of first particle system}

\item{normweights2}{normalized weights of second particle system}

\item{normweights3}{normalized weights of third particle system}

\item{normweights4}{normalized weights of fourth particle system}

\item{ndraws}{number of samples from multinomial distribution}

\item{rand}{common uniform random variables used for the coupled resampling}
}
\value{
A matrix of ancestor indexes of size ndraws x 4
}
\description{
This function performs (4-way) coupled multinomial resampling by maximally coupling chains and 
maximally coupling across levels with coupled residuals using common uniform random variables on each particle system
}
