% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPF.R
\name{CPF}
\alias{CPF}
\title{Conditional Particle Filter}
\usage{
CPF(model, theta, level, observations, nparticles, ref_trajectory = NULL)
}
\arguments{
\item{model}{a list representing a hidden Markov model, e.g. \code{\link{hmm_ornstein_uhlenbeck}}}

\item{theta}{a vector of parameters as input to model functions}

\item{level}{level of the time discretization}

\item{observations}{a matrix of observations of size terminal_time x ydimension}

\item{nparticles}{number of particles}

\item{ref_trajectory}{a matrix of reference trajectory, of size xdimension x statelength; 
if missing, this function runs a standard particle filter}
}
\value{
a matrix containing a new trajectory of size xdimension x statelength
}
\description{
Runs a conditional particle filter
}
