% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm_logistic_diffusion.R
\name{hmm_logistic_diffusion}
\alias{hmm_logistic_diffusion}
\title{Construct hidden Markov model obtained by discretizing logistic diffusion process}
\usage{
hmm_logistic_diffusion(times, sigma0)
}
\arguments{
\item{times}{vector specifying observation times}

\item{sigma0}{diffusivity parameter}
}
\value{
a list with objects such as: 
\code{xdimension} is the dimension of the latent process; 
\code{ydimension} is the dimension of the observation process; 
\code{theta_dimension} is the dimension of the parameter space; 
\code{construct_discretization} outputs a list containing stepsize, nsteps, statelength and obstimes; 
\code{construct_successive_discretization} outputs lists containing stepsize, nsteps, statelength, obstimes for fine and coarse levels, 
and coarsetimes of length statelength_fine indexing time steps of coarse level;
\code{sigma} is the diffusion coefficient of the process; 
\code{rinit} to sample from the initial distribution; 
\code{rtransition} to sample from the Markov transition; 
\code{dtransition} to evaluate the transition density; 
\code{dmeasurement} to evaluate the measurement density; 
\code{functional} is the smoothing function to compute gradient of log-likelihood.
}
\description{
Define a hidden Markov model by discretizing logistic diffusion process.
}
